Update to 1.4.1 release of origin
Includes results of "gofmt -w". Modify the code to work against
origin 1.4.1. The major change here is that scaling must be
done against the deploymentconfig and not the replication
controller.

+++ clusters/services/services.go
@@ -6,31 +6,30 @@ import (
 )
 
 /*
-        {
-            "apiVersion": "v1",
-            "kind": "Service",
-            "metadata": {
-                "labels": {
-                    "name": "spark-master-jpda"
-                },
-                "name": "spark-master-jpda"
-            },
-            "spec": {
-                "ports": [
-                    {
-                        "port": 7077,
-                        "protocol": "TCP",
-                        "targetPort": 7077
-                    }
-                ],
-                "selector": {
-                    "name": "spark-master-jpda"
-                }
-            }
-        },
+   {
+       "apiVersion": "v1",
+       "kind": "Service",
+       "metadata": {
+           "labels": {
+               "name": "spark-master-jpda"
+           },
+           "name": "spark-master-jpda"
+       },
+       "spec": {
+           "ports": [
+               {
+                   "port": 7077,
+                   "protocol": "TCP",
+                   "targetPort": 7077
+               }
+           ],
+           "selector": {
+               "name": "spark-master-jpda"
+           }
+       }
+   },
 
- */
-
+*/
 
 type OService struct {
 	kapi.Service
@@ -86,7 +85,7 @@ type OServicePort struct {
 func ServicePort(port int) *OServicePort {
 	s := OServicePort{}
 	s.ServicePort.Protocol = kapi.ProtocolTCP
-	s.ServicePort.Port = port
+	s.ServicePort.Port = int32(port)
 	return &s
 }
 
@@ -131,4 +130,4 @@ type ServicePort struct {
 	// Default is to auto-allocate a port if the ServiceType of this Service requires one.
 	NodePort int `json:"nodePort"`
 }
- */
+*/
